// safeTech-mobile/frontend/types/types/type.d.ts
declare interface MarkerData {
    id: number | string; // or string, depending on your ID type
    latitude: number;
    longitude: number;

}